api/requirements.txt
# Web Framework
django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==23.0.0

# Database
psycopg2-binary==2.9.10

# Async Task Processing
celery[redis]==5.4.0
redis==5.2.1
kombu==5.4.2
amqp==5.3.1
billiard==4.2.1
vine==5.1.0

# Web Scraping
scrapy==2.11.2
scrapy-djangoitem==1.1.1
parsel==1.9.1
w3lib==2.2.1
itemadapter==0.9.0
protego==0.3.1
queuelib==1.7.0

# HTTP & Security
cryptography==43.0.3
pyopenssl==24.3.0
cffi==1.17.1
pycparser==2.22

# Twisted (required by Scrapy)
twisted==24.11.0
attrs==24.2.0
automat==24.8.1
constantly==23.10.4
hyperlink==21.0.0
incremental==24.7.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
service-identity==24.2.0
zope.interface==7.1.1

# XML/HTML Processing
lxml==5.3.0
cssselect==1.2.0

# Utilities
pytz==2024.2
idna==3.10
pydispatcher==2.0.7
pyhamcrest==2.1.0
six==1.17.0
future==1.0.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-celery==1.1.3
factory-boy==3.3.1

# Note: importlib-metadata removed (not needed for Python 3.8+)
# Note: sqlparse, asgiref, zipp are auto-installed by Django
